Update for damage/health function changes

+++ src/main/java/org/spongepowered/common/event/damage/DamageEventHandler.java
@@ -95,8 +95,9 @@ public class DamageEventHandler {
     }
 
     public static Optional<DamageFunction> createHardHatModifier(EntityLivingBase entityLivingBase,
-                                                                                                          DamageSource damageSource) {
+            DamageSource damageSource) {
         if ((damageSource instanceof FallingBlockDamageSource) && !entityLivingBase.getItemStackFromSlot(EntityEquipmentSlot.HEAD).isEmpty()) {
+            // TODO: direct cause creation: bad bad bad
             DamageModifier modifier = DamageModifier.builder()
                 .cause(
                     Cause.of(EventContext.empty(), ((ItemStack) entityLivingBase.getItemStackFromSlot(EntityEquipmentSlot.HEAD)).createSnapshot()))
@@ -110,7 +111,7 @@ public class DamageEventHandler {
     private static double damageToHandle;
 
     public static Optional<List<DamageFunction>> createArmorModifiers(EntityLivingBase entityLivingBase,
-                                                                                                               DamageSource damageSource, double damage) {
+            DamageSource damageSource, double damage) {
         if (!damageSource.isDamageAbsolute()) {
             damage *= 25;
             net.minecraft.item.ItemStack[] inventory = Iterables.toArray(entityLivingBase.getArmorInventoryList(), net.minecraft.item.ItemStack.class);
@@ -161,6 +162,7 @@ public class DamageEventHandler {
                 };
                 ratio += prop.ratio;
 
+                // TODO: direct cause creation: bad bad bad
                 DamageModifier modifier = DamageModifier.builder()
                     .cause(Cause.of(EventContext.empty(), ((org.spongepowered.api.item.inventory.ItemStack) inventory[prop.slot]).createSnapshot(),
                                     prop, // We need this property to refer to the slot.
@@ -217,10 +219,10 @@ public class DamageEventHandler {
         }
     }
 
-    public static Optional<DamageFunction> createResistanceModifier(EntityLivingBase entityLivingBase,
-                                                                                                             DamageSource damageSource) {
+    public static Optional<DamageFunction> createResistanceModifier(EntityLivingBase entityLivingBase, DamageSource damageSource) {
         if (!damageSource.isDamageAbsolute() && entityLivingBase.isPotionActive(MobEffects.RESISTANCE) && damageSource != DamageSource.OUT_OF_WORLD) {
             PotionEffect effect = ((PotionEffect) entityLivingBase.getActivePotionEffect(MobEffects.RESISTANCE));
+            // TODO: direct cause creation: bad bad bad
             return Optional.of(new DamageFunction(DamageModifier.builder()
                                                .cause(Cause.of(EventContext.empty(), effect))
                                                .type(DamageModifierTypes.DEFENSIVE_POTION_EFFECT)
@@ -283,7 +285,7 @@ public class DamageEventHandler {
                             double magicModifier;
                             if (modifier > 0 && modifier <= 20) {
                                 int j = 25 - modifier;
-                                magicModifier = modifierDamage * (float) j;
+                                magicModifier = modifierDamage * j;
                                 modifierDamage = magicModifier / 25.0F;
                             }
                             return - Math.max(actualDamage - modifierDamage, 0.0D);
@@ -292,6 +294,7 @@ public class DamageEventHandler {
                             first = false;
                         }
 
+                        // TODO: direct cause creation: bad bad bad
                         DamageModifier enchantmentModifier = DamageModifier.builder()
                             .cause(Cause.of(EventContext.empty(), enchantment, snapshot, entityLivingBase))
                             .type(DamageModifierTypes.ARMOR_ENCHANTMENT)
@@ -313,6 +316,7 @@ public class DamageEventHandler {
         if (absorptionAmount > 0) {
             DoubleUnaryOperator function = damage ->
                 -(Math.max(damage - Math.max(damage - absorptionAmount, 0.0F), 0.0F));
+                // TODO: direct cause creation: bad bad bad
             DamageModifier modifier = DamageModifier.builder()
                 .cause(Cause.of(EventContext.empty(), entityLivingBase))
                 .type(DamageModifierTypes.ABSORPTION)
@@ -398,6 +402,7 @@ public class DamageEventHandler {
 
                     final Enchantment enchantment = Enchantment.getEnchantmentByID(j);
                     if (enchantment != null) {
+                        // TODO: direct cause creation: bad bad bad
                         final DamageModifier enchantmentModifier = DamageModifier.builder()
                                 .type(DamageModifierTypes.WEAPON_ENCHANTMENT)
                                 .cause(Cause.of(EventContext.empty(), supplier.get(), enchantment))
@@ -414,6 +419,7 @@ public class DamageEventHandler {
     }
 
     public static DamageFunction provideCriticalAttackTuple(EntityPlayer player) {
+        // TODO: direct cause creation: bad bad bad
         final DamageModifier modifier = DamageModifier.builder()
                 .cause(Cause.of(EventContext.empty(), player))
                 .type(DamageModifierTypes.CRITICAL_HIT)
@@ -436,6 +442,7 @@ public class DamageEventHandler {
 
     public static DamageFunction provideCooldownAttackStrengthFunction(EntityPlayer player,
             float attackStrength) {
+        // TODO: direct cause creation: bad bad bad
         final DamageModifier modifier = DamageModifier.builder()
                 .cause(Cause.of(EventContext.empty(), player))
                 .type(DamageModifierTypes.ATTACK_COOLDOWN)
@@ -448,6 +455,7 @@ public class DamageEventHandler {
 
     public static Optional<DamageFunction> createShieldFunction(EntityLivingBase entity, DamageSource source, float amount) {
         if (entity.isActiveItemStackBlocking() && amount > 0.0 && entity.canBlockDamageSource(source)) {
+            // TODO: direct cause creation: bad bad bad
             final DamageModifier modifier = DamageModifier.builder()
                     .cause(Cause.of(EventContext.empty(), entity, ((ItemStack) entity.getActiveItemStack()).createSnapshot()))
                     .type(DamageModifierTypes.SHIELD)
